Restructure package organization
