+++ app/src/main/java/agency/five/tmdb/ui/components/TabList.kt
@@ -4,7 +4,7 @@ package agency.five.tmdb.ui.components
 import agency.five.tmdb.R
 import agency.five.tmdb.navigation.Screens
 import agency.five.tmdb.ui.theme.Typography
-import agency.five.tmdb.ui.viewmodel.HomeScreenViewModel
+import agency.five.tmdb.viewmodel.HomeScreenViewModel
 import android.util.Log
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.animateDpAsState
