+++ app/src/main/java/agency/five/tmdb/ui/screens/FavoritesScreen.kt
@@ -1,10 +1,10 @@
-package agency.five.tmdb
+package agency.five.tmdb.ui.screens
 
 import agency.five.tmdb.navigation.Screens
 import agency.five.tmdb.ui.components.MovieCard
 import agency.five.tmdb.ui.components.MovieItemViewState
 import agency.five.tmdb.ui.theme.Typography
-import agency.five.tmdb.ui.viewmodel.FavoritesScreenViewModel
+import agency.five.tmdb.viewmodel.FavoritesScreenViewModel
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.GridCells
@@ -24,6 +24,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.navigation.NavController
 import org.koin.androidx.compose.viewModel
+import agency.five.tmdb.R
 
 
 @OptIn(ExperimentalFoundationApi::class)
