+++ app/src/main/java/agency/five/tmdb/viewmodel/DetailsScreenViewModel.kt
@@ -1,13 +1,12 @@
-package agency.five.tmdb.ui.viewmodel
+package agency.five.tmdb.viewmodel
 
-import agency.five.tmdb.Actor
-import agency.five.tmdb.MovieApiImpl
-import agency.five.tmdb.MovieRepository
-import agency.five.tmdb.PersonFunction
+
+import agency.five.tmdb.domain.common.Actor
+import agency.five.tmdb.domain.common.PersonFunction
+import agency.five.tmdb.domain.repointerfaces.MovieRepository
 import agency.five.tmdb.ui.components.MovieItemViewState
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
