+++ app/src/main/java/agency/five/tmdb/viewmodel/FavoritesScreenViewModel.kt
@@ -1,15 +1,12 @@
-package agency.five.tmdb.ui.viewmodel
+package agency.five.tmdb.viewmodel
 
-import agency.five.tmdb.MovieRepository
-import agency.five.tmdb.R
+import agency.five.tmdb.domain.repointerfaces.MovieRepository
 import agency.five.tmdb.ui.components.MovieItemViewState
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
